move some of the ZMQ files around so they are all in the same place and use a target vs a name
