+++ src/helics/network/zmq/ZmqCommsCommon.cpp
@@ -7,7 +7,7 @@ SPDX-License-Identifier: BSD-3-Clause
 #include "ZmqCommsCommon.h"
 
 #include "../NetworkBrokerData.hpp"
-
+#include "cppzmq/zmq.hpp"
 #include <string>
 #include <thread>
