+++ src/helics/shared_api_library/helicsExport.cpp
@@ -22,7 +22,6 @@ SPDX-License-Identifier: BSD-3-Clause
 #include <vector>
 #ifdef ENABLE_ZMQ_CORE
 #    include "../network/zmq/zmqContextManager.h"
-#    include "cppzmq/zmq.hpp"
 #endif
 
 const char* helicsGetVersion(void)
@@ -1001,7 +1000,7 @@ MasterObjectHolder::~MasterObjectHolder()
 {
 #ifdef ENABLE_ZMQ_CORE
     if (ZmqContextManager::setContextToLeakOnDelete()) {
-        ZmqContextManager::getContext().close(); // LCOV_EXCL_LINE
+        ZmqContextManager::closeContext(); // LCOV_EXCL_LINE
     }
 #endif
     helics::LoggingCore::setFastShutdown();
